define interface and function placeholders
