+++ pyrsia_node/src/network/handlers.rs
@@ -14,14 +14,16 @@
    limitations under the License.
 */
 
+use futures::lock::Mutex;
 use libp2p::request_response::ResponseChannel;
 use libp2p::Multiaddr;
 use log::{debug, info};
-use pyrsia::artifact_service;
-use pyrsia::artifact_service::storage::ArtifactStorage;
+use pyrsia::artifact_service::service::ArtifactService;
 use pyrsia::network::artifact_protocol::ArtifactResponse;
 use pyrsia::network::client::{ArtifactType, Client};
 use pyrsia::network::idle_metric_protocol::{IdleMetricResponse, PeerMetrics};
+use pyrsia::peer_metrics;
+use std::sync::Arc;
 
 /// Reach out to another node with the specified address
 pub async fn dial_other_peer(mut p2p_client: Client, to_dial: &Multiaddr) {
@@ -32,8 +34,7 @@ pub async fn dial_other_peer(mut p2p_client: Client, to_dial: &Multiaddr) {
 /// Respond to a RequestArtifact event by getting the artifact
 /// based on the provided artifact type and hash.
 pub async fn handle_request_artifact(
-    mut p2p_client: Client,
-    artifact_storage: ArtifactStorage,
+    artifact_service: Arc<Mutex<ArtifactService>>,
     artifact_type: &ArtifactType,
     artifact_id: &str,
     channel: ResponseChannel<ArtifactResponse>,
@@ -43,10 +44,18 @@ pub async fn handle_request_artifact(
         artifact_type, artifact_id
     );
     let content = match artifact_type {
-        ArtifactType::Artifact => get_artifact(artifact_storage, artifact_id)?,
+        ArtifactType::Artifact => artifact_service
+            .lock()
+            .await
+            .get_artifact_locally(artifact_id)?,
     };
 
-    p2p_client.respond_artifact(content, channel).await
+    artifact_service
+        .lock()
+        .await
+        .p2p_client
+        .respond_artifact(content, channel)
+        .await
 }
 
 //Respond to the IdleMetricRequest event
@@ -54,14 +63,9 @@ pub async fn handle_request_idle_metric(
     mut p2p_client: Client,
     channel: ResponseChannel<IdleMetricResponse>,
 ) -> anyhow::Result<()> {
-    let metric = artifact_service::handlers::get_quality_metric();
+    let metric = peer_metrics::metrics::get_quality_metric();
     let peer_metrics: PeerMetrics = PeerMetrics {
         idle_metric: metric.to_le_bytes(),
     };
     p2p_client.respond_idle_metric(peer_metrics, channel).await
-}
-
-/// Get the artifact with the provided hash from the artifact manager.
-fn get_artifact(artifact_storage: ArtifactStorage, artifact_id: &str) -> anyhow::Result<Vec<u8>> {
-    artifact_service::handlers::get_artifact_locally(&artifact_storage, artifact_id)
 }
