+++ src/artifact_service/service.rs
@@ -14,205 +14,396 @@
    limitations under the License.
 */
 
-use anyhow::anyhow;
+use super::hashing::{Hash, HashAlgorithm};
+use super::storage::ArtifactStorage;
+use crate::network::client::{ArtifactType, Client};
+use crate::transparency_log::log::{TransparencyLog, TransparencyLogError};
+use anyhow::{bail, Context};
+use libp2p::PeerId;
+use log::info;
+use multihash::Hasher;
 use serde::{Deserialize, Serialize};
-use sha2::{Digest, Sha256, Sha512};
-use std::fmt::{Display, Formatter};
-use strum_macros::{EnumIter, EnumString};
+use std::fs::File;
+use std::io::{BufReader, Read};
+use std::path::Path;
+use std::str;
 
-// We will provide implementations of this trait for each hash algorithm that we support.
-pub trait Digester {
-    fn update_hash(&mut self, input: &[u8]);
-
-    fn finalize_hash(&mut self, hash_buffer: &mut [u8]);
-
-    fn hash_size_in_bytes(&self) -> usize;
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize, strum_macros::Display)]
+pub enum PackageType {
+    Docker,
+    Maven2,
 }
 
-impl Digester for Sha256 {
-    fn update_hash(&mut self, input: &[u8]) {
-        self.update(input);
-    }
-
-    fn finalize_hash(&mut self, hash_buffer: &mut [u8]) {
-        hash_buffer.clone_from_slice(self.clone().finalize().as_slice());
-    }
-
-    fn hash_size_in_bytes(&self) -> usize {
-        256 / 8
-    }
+pub struct ArtifactService {
+    pub artifact_storage: ArtifactStorage,
+    pub transparency_log: TransparencyLog,
+    pub p2p_client: Client,
 }
 
-impl Digester for Sha512 {
-    fn update_hash(&mut self, input: &[u8]) {
-        self.update(input);
+impl ArtifactService {
+    pub fn new<P: AsRef<Path>>(artifact_path: P, p2p_client: Client) -> anyhow::Result<Self> {
+        let artifact_storage = ArtifactStorage::new(&artifact_path)?;
+        let transparency_log = TransparencyLog::new(&artifact_path)?;
+        Ok(ArtifactService {
+            artifact_storage,
+            transparency_log,
+            p2p_client,
+        })
     }
 
-    fn finalize_hash(&mut self, hash_buffer: &mut [u8]) {
-        hash_buffer.clone_from_slice(self.clone().finalize().as_slice());
+    pub fn request_build(&self, _package_type: PackageType, _package_type_id: &str) {}
+
+    pub async fn get_artifact(
+        &mut self,
+        package_type: PackageType,
+        package_type_id: &str,
+    ) -> anyhow::Result<Vec<u8>> {
+        let transaction = self
+            .transparency_log
+            .get_artifact(&package_type, package_type_id)?;
+
+        let blob_content = match self.get_artifact_locally(&transaction.hash) {
+            Ok(blob_content) => Ok(blob_content),
+            Err(_) => self.get_artifact_from_peers(&transaction.hash).await,
+        }?;
+
+        self.verify_artifact(&package_type, package_type_id, &blob_content)
+            .await?;
+
+        Ok(blob_content)
     }
 
-    fn hash_size_in_bytes(&self) -> usize {
-        512 / 8
+    pub fn get_artifact_locally(&mut self, artifact_id: &str) -> Result<Vec<u8>, anyhow::Error> {
+        let decoded_hash = hex::decode(artifact_id)?;
+        let hash: Hash = Hash::new(HashAlgorithm::SHA256, &decoded_hash)?;
+        let result = self.artifact_storage.pull_artifact(&hash)?;
+        let mut buf_reader: BufReader<File> = BufReader::new(result);
+        let mut blob_content = Vec::new();
+        buf_reader.read_to_end(&mut blob_content)?;
+        Ok(blob_content)
     }
-}
 
-/// The types of hash algorithms that the artifact service supports
-#[derive(EnumIter, Clone, Debug, PartialEq, EnumString, Serialize, Deserialize)]
-pub enum HashAlgorithm {
-    SHA256,
-    SHA512,
-}
+    async fn get_artifact_from_peers(
+        &mut self,
+        artifact_id: &str,
+    ) -> Result<Vec<u8>, anyhow::Error> {
+        let providers = self
+            .p2p_client
+            .list_providers(ArtifactType::Artifact, artifact_id.into())
+            .await?;
 
-impl HashAlgorithm {
-    pub fn digest_factory(&self) -> Box<dyn Digester> {
-        match self {
-            HashAlgorithm::SHA256 => Box::new(Sha256::new()),
-            HashAlgorithm::SHA512 => Box::new(Sha512::new()),
+        match self.p2p_client.get_idle_peer(providers).await? {
+            Some(peer) => self.get_artifact_from_peer(&peer, artifact_id).await,
+            None => {
+                bail!(
+                    "Artifact with id {} is not available on the p2p network.",
+                    artifact_id
+                )
+            }
         }
     }
 
-    /// Translate a HashAlgorithm to a string.
-    pub fn hash_algorithm_to_str(&self) -> &'static str {
-        match self {
-            HashAlgorithm::SHA256 => "SHA256",
-            HashAlgorithm::SHA512 => "SHA512",
-        }
+    async fn get_artifact_from_peer(
+        &mut self,
+        peer_id: &PeerId,
+        artifact_id: &str,
+    ) -> Result<Vec<u8>, anyhow::Error> {
+        let artifact = self
+            .p2p_client
+            .request_artifact(peer_id, ArtifactType::Artifact, artifact_id.into())
+            .await?;
+
+        let decoded_hash = hex::decode(artifact_id)?;
+        let hash: Hash = Hash::new(HashAlgorithm::SHA256, &decoded_hash)?;
+        let cursor = Box::new(std::io::Cursor::new(artifact));
+        self.put_artifact(&hash, cursor)?;
+        self.get_artifact_locally(artifact_id)
     }
 
-    fn hash_length_in_bytes(&self) -> usize {
-        match self {
-            HashAlgorithm::SHA256 => 256 / 8,
-            HashAlgorithm::SHA512 => 512 / 8,
-        }
-    }
-}
+    async fn verify_artifact(
+        &mut self,
+        package_type: &PackageType,
+        package_type_id: &str,
+        blob_content: &[u8],
+    ) -> Result<(), TransparencyLogError> {
+        let mut sha256 = multihash::Sha2_256::default();
+        sha256.update(blob_content);
+        let calculated_hash = hex::encode(sha256.finalize());
 
-impl Display for HashAlgorithm {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        f.write_str(HashAlgorithm::hash_algorithm_to_str(self))
-    }
-}
-
-#[derive(PartialEq, Debug)]
-pub struct Hash {
-    pub algorithm: HashAlgorithm,
-    pub bytes: Vec<u8>,
-}
-
-impl Hash {
-    pub fn new(algorithm: HashAlgorithm, bytes: &[u8]) -> Result<Self, anyhow::Error> {
-        let expected_length: usize = algorithm.hash_length_in_bytes();
-        if bytes.len() == expected_length {
-            Ok(Hash {
-                algorithm,
-                bytes: bytes.to_vec(),
-            })
+        let transaction = self
+            .transparency_log
+            .get_artifact(package_type, package_type_id)?;
+        if transaction.hash == calculated_hash {
+            Ok(())
         } else {
-            Err(anyhow!(format!("The hash value does not have the correct length for the algorithm. The expected length is {} bytes, but the length of the supplied hash is {}.", expected_length, bytes.len())))
+            Err(TransparencyLogError::InvalidHash {
+                id: package_type_id.to_string(),
+                invalid_hash: calculated_hash,
+                actual_hash: transaction.hash,
+            })
         }
     }
-}
 
-impl Display for Hash {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!(
-            "{}:{}",
-            self.algorithm.hash_algorithm_to_str(),
-            hex::encode(&self.bytes)
-        ))
+    //put_artifact: given artifact_hash(artifactName) & artifact_path push artifact to artifact_manager
+    //              and returns the boolean as true or false if it was able to create or not
+    fn put_artifact(
+        &self,
+        artifact_hash: &Hash,
+        art_reader: Box<dyn Read>,
+    ) -> Result<(), anyhow::Error> {
+        info!("put_artifact hash: {}", artifact_hash);
+        let mut buf_reader = BufReader::new(art_reader);
+        self.artifact_storage
+            .push_artifact(&mut buf_reader, artifact_hash)
+            .context("Error from put_artifact")
     }
 }
 
 #[cfg(test)]
 mod tests {
-    pub use super::*;
-    use strum::IntoEnumIterator;
+    use super::*;
+    use crate::network::client::command::Command;
+    use crate::network::idle_metric_protocol::PeerMetrics;
+    use crate::transparency_log::log::TransparencyLogError;
+    use crate::util::test_util;
+    use anyhow::Context;
+    use futures::channel::mpsc;
+    use futures::executor;
+    use futures::prelude::*;
+    use libp2p::identity::Keypair;
+    use sha2::{Digest, Sha256};
+    use std::collections::HashSet;
+    use std::env;
+    use std::fs::File;
+    use std::path::PathBuf;
 
-    const TEST_ARTIFACT_HASH_256: [u8; 32] = [
-        0x6b, 0x29, 0xf2, 0xf1, 0xe5, 0x02, 0x4c, 0x41, 0x95, 0x06, 0xe9, 0x50, 0x3e, 0x02, 0x4b,
-        0x3d, 0x8a, 0x5a, 0x08, 0xb6, 0xf6, 0xd5, 0x5b, 0x68, 0x88, 0x66, 0x79, 0x52, 0xd1, 0x04,
-        0x15, 0x54,
+    const VALID_ARTIFACT_HASH: [u8; 32] = [
+        0x86, 0x5c, 0x8d, 0x98, 0x8b, 0xe4, 0x66, 0x9f, 0x3e, 0x48, 0xf7, 0x3b, 0x98, 0xf9, 0xbc,
+        0x25, 0x7, 0xbe, 0x2, 0x46, 0xea, 0x35, 0xe0, 0x9, 0x8c, 0xf6, 0x5, 0x4d, 0x36, 0x44, 0xc1,
+        0x4f,
     ];
-    const TEST_ARTIFACT_HASH_512: [u8; 64] = [
-        0x6b, 0x29, 0xf2, 0xf1, 0xe5, 0x02, 0x4c, 0x41, 0x95, 0x06, 0xe9, 0x50, 0x3e, 0x02, 0x4b,
-        0x3d, 0x8a, 0x5a, 0x08, 0xb6, 0xf6, 0xd5, 0x5b, 0x68, 0x88, 0x66, 0x79, 0x52, 0xd1, 0x04,
-        0x15, 0x54, 0x83, 0x74, 0x5a, 0xc0, 0x84, 0xfe, 0xf2, 0x12, 0x29, 0xd6, 0x57, 0x2c, 0xd4,
-        0x14, 0xf9, 0xb2, 0xa4, 0x82, 0x06, 0xd6, 0x47, 0x62, 0xc5, 0x26, 0x81, 0x11, 0xd1, 0xc4,
-        0x7a, 0x87, 0x4e, 0x71,
-    ];
 
-    const SHA256_HASH_ENCODED: &str =
-        "5e6009f8ce7a159884aa5e5132ce8c84fefc979f237a0bce4652f90bc77e5591";
-    const SHA512_HASH_ENCODED: &str = "838d2542932c2545f222a4daf74e0e1dc1bd76ce5742b4e3a92aaff2e28b038adf50e0bbdfe6da50ff4fc19f8a23a77ce8fd28a38456b33d43a62b3c86978954";
+    #[tokio::test]
+    async fn test_put_and_get_artifact() {
+        let tmp_dir = test_util::tests::setup();
 
-    #[test]
-    pub fn test_digester_length() {
-        for algorithm in HashAlgorithm::iter() {
-            let digester = algorithm.digest_factory();
+        let (sender, _receiver) = mpsc::channel(1);
+        let p2p_client = Client {
+            sender,
+            local_peer_id: Keypair::generate_ed25519().public().to_peer_id(),
+        };
 
-            assert_eq!(
-                digester.hash_size_in_bytes(),
-                algorithm.hash_length_in_bytes()
-            );
-        }
+        let mut artifact_service = ArtifactService::new(&tmp_dir, p2p_client).unwrap();
+
+        let package_type = PackageType::Docker;
+        let artifact_id = "an_artifact_id";
+        artifact_service
+            .transparency_log
+            .add_artifact(
+                &package_type,
+                artifact_id,
+                &hex::encode(VALID_ARTIFACT_HASH),
+            )
+            .unwrap();
+
+        let hash = Hash::new(HashAlgorithm::SHA256, &VALID_ARTIFACT_HASH).unwrap();
+        //put the artifact
+        artifact_service
+            .put_artifact(&hash, Box::new(get_file_reader().unwrap()))
+            .context("Error from put_artifact")
+            .unwrap();
+
+        // pull artifact
+        let future = async {
+            artifact_service
+                .get_artifact(package_type, artifact_id)
+                .await
+                .context("Error from get_artifact")
+        };
+        let file = executor::block_on(future).unwrap();
+
+        //validate pulled artifact with the actual data
+        let mut s = String::new();
+        get_file_reader().unwrap().read_to_string(&mut s).unwrap();
+
+        let s1 = match str::from_utf8(file.as_slice()) {
+            Ok(v) => v,
+            Err(e) => panic!("Invalid UTF-8 sequence: {}", e),
+        };
+        assert_eq!(s, s1);
+
+        test_util::tests::teardown(tmp_dir);
     }
 
-    #[test]
-    pub fn test_digester_256() {
-        let mut digester = HashAlgorithm::SHA256.digest_factory();
-        digester.update_hash(&TEST_ARTIFACT_HASH_256);
-        let mut hash_buffer = [0; 32];
-        digester.finalize_hash(&mut hash_buffer);
-        assert_eq!(hex::encode(hash_buffer), SHA256_HASH_ENCODED);
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_get_from_peers() {
+        let tmp_dir = test_util::tests::setup();
+
+        let (sender, mut receiver) = mpsc::channel(1);
+        let local_peer_id = Keypair::generate_ed25519().public().to_peer_id();
+        let p2p_client = Client {
+            sender,
+            local_peer_id,
+        };
+
+        let mut artifact_service = ArtifactService::new(&tmp_dir, p2p_client).unwrap();
+
+        tokio::spawn(async move {
+            loop {
+                match receiver.next().await {
+                    Some(Command::ListProviders { artifact_type: _artifact_type, artifact_hash: _artifact_hash, sender }) => {
+                        let mut set = HashSet::new();
+                        set.insert(local_peer_id);
+                        let _ = sender.send(set);
+                    },
+                    Some(Command::RequestIdleMetric { peer: _peer, sender }) => {
+                        let _ = sender.send(Ok(PeerMetrics {
+                            idle_metric: (0.1_f64).to_le_bytes()
+                        }));
+                    },
+                    Some(Command::RequestArtifact { artifact_type: _artifact_type, artifact_hash: _artifact_hash, peer: _peer, sender }) => {
+                        let _ = sender.send(Ok(b"SAMPLE_DATA".to_vec()));
+                    },
+                    _ => panic!("Command must match Command::ListProviders, Command::RequestIdleMetric, Command::RequestArtifact"),
+                }
+            }
+        });
+
+        let mut hasher = Sha256::new();
+        hasher.update(b"SAMPLE_DATA");
+        let hash_bytes = hasher.finalize();
+        let artifact_id = hex::encode(hash_bytes);
+
+        let result = executor::block_on(async {
+            artifact_service.get_artifact_from_peers(&artifact_id).await
+        });
+        assert!(result.is_ok());
+
+        test_util::tests::teardown(tmp_dir);
     }
 
-    #[test]
-    pub fn test_digester_512() {
-        let mut digester = HashAlgorithm::SHA512.digest_factory();
-        digester.update_hash(&TEST_ARTIFACT_HASH_256);
-        let mut hash_buffer = [0; 64];
-        digester.finalize_hash(&mut hash_buffer);
-        assert_eq!(hex::encode(hash_buffer), SHA512_HASH_ENCODED);
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_get_from_peers_with_no_providers() {
+        let tmp_dir = test_util::tests::setup();
+
+        let (sender, mut receiver) = mpsc::channel(1);
+        let peer_id = Keypair::generate_ed25519().public().to_peer_id();
+        let p2p_client = Client {
+            sender,
+            local_peer_id: peer_id,
+        };
+
+        let mut artifact_service = ArtifactService::new(&tmp_dir, p2p_client).unwrap();
+
+        tokio::spawn(async move {
+            futures::select! {
+                command = receiver.next() => {
+                    match command {
+                        Some(Command::ListProviders { artifact_type: _artifact_type, artifact_hash: _artifact_hash, sender }) => {
+                            let _ = sender.send(Default::default());
+                        },
+                        _ => panic!("Command must match Command::ListProviders"),
+                    }
+                }
+            }
+        });
+
+        let mut hasher = Sha256::new();
+        hasher.update(b"SAMPLE_DATA");
+        let hash_bytes = hasher.finalize();
+        let artifact_id = hex::encode(hash_bytes);
+
+        let result = executor::block_on(async {
+            artifact_service.get_artifact_from_peers(&artifact_id).await
+        });
+        assert!(result.is_err());
+
+        test_util::tests::teardown(tmp_dir);
     }
 
-    #[test]
-    pub fn test_hash_new() {
-        let hash = Hash::new(HashAlgorithm::SHA256, &TEST_ARTIFACT_HASH_256).unwrap();
+    #[tokio::test]
+    async fn test_verify_artifact_succeeds_when_hashes_same() {
+        let tmp_dir = test_util::tests::setup();
 
-        assert_eq!(hash.algorithm, HashAlgorithm::SHA256);
-        assert_eq!(hash.bytes, TEST_ARTIFACT_HASH_256);
+        let (sender, _receiver) = mpsc::channel(1);
+        let local_peer_id = Keypair::generate_ed25519().public().to_peer_id();
+        let p2p_client = Client {
+            sender,
+            local_peer_id,
+        };
+
+        let mut artifact_service = ArtifactService::new(&tmp_dir, p2p_client).unwrap();
+
+        let mut hasher1 = Sha256::new();
+        hasher1.update(b"SAMPLE_DATA");
+        let random_hash = hex::encode(hasher1.finalize());
+
+        let package_type = PackageType::Docker;
+        let package_type_id = "package_type_id";
+        artifact_service
+            .transparency_log
+            .add_artifact(&package_type, package_type_id, &random_hash)
+            .unwrap();
+
+        let result = artifact_service
+            .verify_artifact(&package_type, package_type_id, b"SAMPLE_DATA")
+            .await;
+        assert!(result.is_ok());
+
+        test_util::tests::teardown(tmp_dir);
     }
 
-    #[test]
-    pub fn test_hash_256_display() {
-        let hash = Hash::new(HashAlgorithm::SHA256, &TEST_ARTIFACT_HASH_256).unwrap();
-        let display = format!("{}", hash);
-        let to_string = format!(
-            "{}:{}",
-            hash.algorithm.hash_algorithm_to_str(),
-            hex::encode(hash.bytes)
+    #[tokio::test]
+    async fn test_verify_artifact_fails_when_hashes_differ() {
+        let tmp_dir = test_util::tests::setup();
+
+        let (sender, _receiver) = mpsc::channel(1);
+        let local_peer_id = Keypair::generate_ed25519().public().to_peer_id();
+        let p2p_client = Client {
+            sender,
+            local_peer_id,
+        };
+
+        let mut artifact_service = ArtifactService::new(&tmp_dir, p2p_client).unwrap();
+
+        let mut hasher1 = Sha256::new();
+        hasher1.update(b"SAMPLE_DATA");
+        let random_hash = hex::encode(hasher1.finalize());
+
+        let mut hasher2 = Sha256::new();
+        hasher2.update(b"OTHER_SAMPLE_DATA");
+        let random_other_hash = hex::encode(hasher2.finalize());
+
+        let package_type = PackageType::Docker;
+        let package_type_id = "package_type_id";
+        artifact_service
+            .transparency_log
+            .add_artifact(&package_type, package_type_id, &random_hash)
+            .unwrap();
+
+        let result = artifact_service
+            .verify_artifact(&package_type, package_type_id, b"OTHER_SAMPLE_DATA")
+            .await;
+        assert!(result.is_err());
+        assert_eq!(
+            result.unwrap_err(),
+            TransparencyLogError::InvalidHash {
+                id: package_type_id.to_string(),
+                invalid_hash: random_other_hash,
+                actual_hash: random_hash
+            }
         );
-        assert_eq!(display, to_string);
+
+        test_util::tests::teardown(tmp_dir);
     }
 
-    #[test]
-    pub fn test_hash_512_display() {
-        let hash = Hash::new(HashAlgorithm::SHA512, &TEST_ARTIFACT_HASH_512).unwrap();
-        let display = format!("{}", hash);
-        let to_string = format!(
-            "{}:{}",
-            hash.algorithm.hash_algorithm_to_str(),
-            hex::encode(hash.bytes)
-        );
-        assert_eq!(display, to_string);
-    }
+    fn get_file_reader() -> Result<File, anyhow::Error> {
+        // test artifact file in resources/test dir
+        let mut curr_dir = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+        curr_dir.push("tests/resources/artifact_test.json");
 
-    #[test]
-    pub fn hash_length_does_not_match_algorithm_test() {
-        assert!(
-            Hash::new(HashAlgorithm::SHA512, &[0u8; 7]).is_err(),
-            "A 56 bit hash value for SHA512 should be an error"
-        )
+        let path = String::from(curr_dir.to_string_lossy());
+        let reader = File::open(path.as_str()).unwrap();
+        Ok(reader)
     }
 }
