+++ src/docker/v2/handlers/blobs.rs
@@ -14,11 +14,8 @@
    limitations under the License.
 */
 
-use crate::artifact_service::handlers::get_artifact;
-use crate::artifact_service::storage::ArtifactStorage;
+use crate::artifact_service::service::{ArtifactService, PackageType};
 use crate::docker::error_util::{RegistryError, RegistryErrorCode};
-use crate::network::client::Client;
-use crate::transparency_log::log::TransparencyLog;
 use futures::lock::Mutex;
 use log::debug;
 use std::result::Result;
@@ -26,25 +23,21 @@ use std::sync::Arc;
 use warp::{http::StatusCode, Rejection, Reply};
 
 pub async fn handle_get_blobs(
-    transparency_log: Arc<Mutex<TransparencyLog>>,
-    p2p_client: Client,
-    artifact_storage: ArtifactStorage,
+    artifact_service: Arc<Mutex<ArtifactService>>,
     hash: String,
 ) -> Result<impl Reply, Rejection> {
     debug!("Getting blob with hash : {:?}", hash);
 
-    let blob_content = get_artifact(
-        transparency_log,
-        p2p_client,
-        &artifact_storage,
-        &get_namespace_specific_id(&hash),
-    )
-    .await
-    .map_err(|_| {
-        warp::reject::custom(RegistryError {
-            code: RegistryErrorCode::BlobUnknown,
-        })
-    })?;
+    let blob_content = artifact_service
+        .lock()
+        .await
+        .get_artifact(PackageType::Docker, &hash)
+        .await
+        .map_err(|_| {
+            warp::reject::custom(RegistryError {
+                code: RegistryErrorCode::BlobUnknown,
+            })
+        })?;
 
     Ok(warp::http::response::Builder::new()
         .header("Content-Type", "application/octet-stream")
@@ -53,14 +46,12 @@ pub async fn handle_get_blobs(
         .unwrap())
 }
 
-fn get_namespace_specific_id(hash: &str) -> String {
-    format!("DOCKER::BLOB::{}", hash)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::artifact_service::service::{Hash, HashAlgorithm};
+    use crate::artifact_service::hashing::{Hash, HashAlgorithm};
+    use crate::artifact_service::storage::ArtifactStorage;
+    use crate::network::client::Client;
     use crate::util::test_util;
     use anyhow::Context;
     use futures::channel::mpsc;
@@ -76,29 +67,13 @@ mod tests {
         0x4f,
     ];
 
-    #[test]
-    fn test_get_namespace_specific_id() {
-        let hash = "hash";
-
-        assert_eq!(
-            get_namespace_specific_id(hash),
-            format!("DOCKER::BLOB::{}", hash)
-        );
-    }
-
     #[tokio::test]
     async fn test_handle_get_blobs_unknown_in_artifact_service() {
         let tmp_dir = test_util::tests::setup();
 
         let hash = "7300a197d7deb39371d4683d60f60f2fbbfd7541837ceb2278c12014e94e657b";
-        let namespace_specific_id = format!("DOCKER::BLOB::{}", hash);
-
-        let transparency_log = Arc::new(Mutex::new(TransparencyLog::new(&tmp_dir).unwrap()));
-        transparency_log
-            .lock()
-            .await
-            .add_artifact(&namespace_specific_id, hash)
-            .unwrap();
+        let package_type = PackageType::Docker;
+        let package_type_id = hash;
 
         let (sender, _) = mpsc::channel(1);
         let p2p_client = Client {
@@ -106,15 +81,16 @@ mod tests {
             local_peer_id: Keypair::generate_ed25519().public().to_peer_id(),
         };
 
-        let artifact_storage = ArtifactStorage::new(&tmp_dir).unwrap();
+        let mut artifact_service =
+            ArtifactService::new(&tmp_dir, p2p_client).expect("Creating ArtifactService failed");
 
-        let result = handle_get_blobs(
-            transparency_log,
-            p2p_client,
-            artifact_storage,
-            hash.to_string(),
-        )
-        .await;
+        artifact_service
+            .transparency_log
+            .add_artifact(&package_type, &package_type_id, hash)
+            .unwrap();
+
+        let result =
+            handle_get_blobs(Arc::new(Mutex::new(artifact_service)), hash.to_string()).await;
 
         assert!(result.is_err());
         let rejection = result.err().unwrap();
@@ -134,14 +110,8 @@ mod tests {
         let tmp_dir = test_util::tests::setup();
 
         let hash = "865c8d988be4669f3e48f73b98f9bc2507be0246ea35e0098cf6054d3644c14f";
-        let namespace_specific_id = format!("DOCKER::BLOB::{}", hash);
-
-        let transparency_log = Arc::new(Mutex::new(TransparencyLog::new(&tmp_dir).unwrap()));
-        transparency_log
-            .lock()
-            .await
-            .add_artifact(&namespace_specific_id, hash)
-            .unwrap();
+        let package_type = PackageType::Docker;
+        let package_type_id = hash;
 
         let (sender, _) = mpsc::channel(1);
         let p2p_client = Client {
@@ -149,16 +119,16 @@ mod tests {
             local_peer_id: Keypair::generate_ed25519().public().to_peer_id(),
         };
 
-        let artifact_storage = ArtifactStorage::new(&tmp_dir).unwrap();
-        create_artifact(&artifact_storage).unwrap();
+        let mut artifact_service =
+            ArtifactService::new(&tmp_dir, p2p_client).expect("Creating ArtifactService failed");
+        artifact_service
+            .transparency_log
+            .add_artifact(&package_type, &package_type_id, hash)
+            .unwrap();
+        create_artifact(&artifact_service.artifact_storage).unwrap();
 
-        let result = handle_get_blobs(
-            transparency_log,
-            p2p_client,
-            artifact_storage,
-            hash.to_string(),
-        )
-        .await;
+        let result =
+            handle_get_blobs(Arc::new(Mutex::new(artifact_service)), hash.to_string()).await;
 
         assert!(result.is_ok());
