+++ src/docker/v2/routes.rs
@@ -15,18 +15,16 @@
    limitations under the License.
 */
 
-use crate::artifact_service::storage::ArtifactStorage;
-use crate::docker::v2::handlers::{blobs::handle_get_blobs, manifests::fetch_manifest};
-use crate::network::client::Client;
-use crate::transparency_log::log::TransparencyLog;
+use crate::artifact_service::service::ArtifactService;
+
+use super::handlers::blobs::*;
+use super::handlers::manifests::*;
 use futures::lock::Mutex;
 use std::sync::Arc;
 use warp::Filter;
 
 pub fn make_docker_routes(
-    transparency_log: TransparencyLog,
-    p2p_client: Client,
-    artifact_storage: ArtifactStorage,
+    artifact_service: Arc<Mutex<ArtifactService>>,
 ) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
     let empty_json = "{}";
     let v2_base = warp::path("v2")
@@ -42,34 +40,18 @@ pub fn make_docker_routes(
             "application/json",
         ));
 
-    let transparency_log_fetch_manifest = Arc::new(Mutex::new(transparency_log));
-    let transparency_log_get_blobs = transparency_log_fetch_manifest.clone();
-    let p2p_client_fetch_manifest = p2p_client.clone();
-    let artifact_storage_fetch_manifest = artifact_storage.clone();
+    let artifact_service_get_manifest = artifact_service.clone();
 
     let v2_manifests = warp::path!("v2" / "library" / String / "manifests" / String)
         .and(warp::get().or(warp::head()).unify())
         .and_then(move |name, tag| {
-            fetch_manifest(
-                transparency_log_fetch_manifest.clone(),
-                p2p_client_fetch_manifest.clone(),
-                artifact_storage_fetch_manifest.clone(),
-                name,
-                tag,
-            )
+            fetch_manifest(artifact_service_get_manifest.clone(), name, tag)
         });
 
     let v2_blobs = warp::path!("v2" / "library" / String / "blobs" / String)
         .and(warp::get().or(warp::head()).unify())
         .and(warp::path::end())
-        .and_then(move |_name, hash| {
-            handle_get_blobs(
-                transparency_log_get_blobs.clone(),
-                p2p_client.clone(),
-                artifact_storage.clone(),
-                hash,
-            )
-        });
+        .and_then(move |_name, hash| handle_get_blobs(artifact_service.clone(), hash));
 
     warp::any().and(v2_base.or(v2_manifests).or(v2_blobs))
 }
