+++ src/java/maven2/handlers/maven_artifacts.rs
@@ -14,11 +14,8 @@
    limitations under the License.
 */
 
-use crate::artifact_service::handlers::get_artifact;
-use crate::artifact_service::storage::ArtifactStorage;
+use crate::artifact_service::service::{ArtifactService, PackageType};
 use crate::docker::error_util::{RegistryError, RegistryErrorCode};
-use crate::network::client::Client;
-use crate::transparency_log::log::TransparencyLog;
 use anyhow::bail;
 use futures::lock::Mutex;
 use log::debug;
@@ -26,34 +23,30 @@ use std::sync::Arc;
 use warp::{http::StatusCode, Rejection, Reply};
 
 pub async fn handle_get_maven_artifact(
-    transparency_log: Arc<Mutex<TransparencyLog>>,
-    p2p_client: Client,
-    artifact_storage: ArtifactStorage,
+    artifact_service: Arc<Mutex<ArtifactService>>,
     full_path: String,
 ) -> Result<impl Reply, Rejection> {
     debug!("Requesting maven artifact: {}", full_path);
-    let namespace_specific_id = get_namespace_specific_id(&full_path).map_err(|err| {
-        debug!("Error getting namespace id for artifact: {:?}", err);
+    let package_type_id = get_package_type_id(&full_path).map_err(|err| {
+        debug!("Error getting package type id for artifact: {:?}", err);
         warp::reject::custom(RegistryError {
             code: RegistryErrorCode::Unknown(err.to_string()),
         })
     })?;
 
     // request artifact
-    debug!("Requesting artifact for id {}", &namespace_specific_id);
-    let artifact_content = get_artifact(
-        transparency_log,
-        p2p_client,
-        &artifact_storage,
-        &namespace_specific_id,
-    )
-    .await
-    .map_err(|err| {
-        debug!("Error retrieving artifact: {:?}", err);
-        warp::reject::custom(RegistryError {
-            code: RegistryErrorCode::Unknown(err.to_string()),
-        })
-    })?;
+    debug!("Requesting artifact for id {}", package_type_id);
+    let artifact_content = artifact_service
+        .lock()
+        .await
+        .get_artifact(PackageType::Maven2, &package_type_id)
+        .await
+        .map_err(|err| {
+            debug!("Error retrieving artifact: {:?}", err);
+            warp::reject::custom(RegistryError {
+                code: RegistryErrorCode::Unknown(err.to_string()),
+            })
+        })?;
 
     Ok(warp::http::response::Builder::new()
         .header("Content-Type", "application/octet-stream")
@@ -62,7 +55,7 @@ pub async fn handle_get_maven_artifact(
         .unwrap())
 }
 
-fn get_namespace_specific_id(full_path: &str) -> Result<String, anyhow::Error> {
+fn get_package_type_id(full_path: &str) -> Result<String, anyhow::Error> {
     // maven coordinates like "com.company:test:1.0" will produce a request
     // like: "GET /maven2/com/company/test/1.0/test-1.0.jar"
 
@@ -77,7 +70,7 @@ fn get_namespace_specific_id(full_path: &str) -> Result<String, anyhow::Error> {
     let group_id = pieces.join(".");
 
     Ok(format!(
-        "MAVEN2/FILE/{}/{}/{}/{}",
+        "{}/{}/{}/{}",
         group_id, artifact_id, version, file_name
     ))
 }
@@ -85,7 +78,9 @@ fn get_namespace_specific_id(full_path: &str) -> Result<String, anyhow::Error> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::artifact_service::service::{Hash, HashAlgorithm};
+    use crate::artifact_service::hashing::{Hash, HashAlgorithm};
+    use crate::artifact_service::storage::ArtifactStorage;
+    use crate::network::client::Client;
     use crate::util::test_util;
     use anyhow::Context;
     use futures::channel::mpsc;
@@ -98,31 +93,24 @@ mod tests {
         "e11c16ff163ccc1efe01d2696c626891560fa82123601a5ff196d97b6ab156da";
     const VALID_FULL_PATH: &str = "/maven2/test/test/1.0/test-1.0.jar";
     const INVALID_FULL_PATH: &str = "/maven2/test/1.0/test-1.0.jar";
-    const VALID_MAVEN_ID: &str = "MAVEN2/FILE/test/test/1.0/test-1.0.jar";
+    const VALID_MAVEN_ID: &str = "test/test/1.0/test-1.0.jar";
 
     #[test]
-    fn get_namespace_specific_id_test() {
+    fn get_package_type_id_test() {
         assert_eq!(
-            get_namespace_specific_id(VALID_FULL_PATH).unwrap(),
+            get_package_type_id(VALID_FULL_PATH).unwrap(),
             VALID_MAVEN_ID
         );
     }
 
     #[test]
-    fn get_namespace_specific_id_with_invalid_path_test() {
-        assert!(get_namespace_specific_id(INVALID_FULL_PATH).is_err());
+    fn get_package_type_id_with_invalid_path_test() {
+        assert!(get_package_type_id(INVALID_FULL_PATH).is_err());
     }
 
     #[tokio::test]
     async fn handle_get_maven_artifact_test() {
         let tmp_dir = test_util::tests::setup();
-
-        let transparency_log = Arc::new(Mutex::new(TransparencyLog::new(&tmp_dir).unwrap()));
-        transparency_log
-            .lock()
-            .await
-            .add_artifact(VALID_MAVEN_ID, VALID_ARTIFACT_HASH)
-            .unwrap();
 
         let (sender, _) = mpsc::channel(1);
         let p2p_client = Client {
@@ -130,13 +118,17 @@ mod tests {
             local_peer_id: Keypair::generate_ed25519().public().to_peer_id(),
         };
 
-        let artifact_storage = ArtifactStorage::new(&tmp_dir).unwrap();
-        create_artifact(&artifact_storage).unwrap();
+        let mut artifact_service =
+            ArtifactService::new(&tmp_dir, p2p_client).expect("Creating ArtifactService failed");
+
+        artifact_service
+            .transparency_log
+            .add_artifact(&PackageType::Maven2, VALID_MAVEN_ID, VALID_ARTIFACT_HASH)
+            .unwrap();
+        create_artifact(&artifact_service.artifact_storage).unwrap();
 
         let result = handle_get_maven_artifact(
-            transparency_log,
-            p2p_client,
-            artifact_storage,
+            Arc::new(Mutex::new(artifact_service)),
             VALID_FULL_PATH.to_string(),
         )
         .await;
