+++ src/lib.rs
@@ -17,11 +17,13 @@
 #![allow(mixed_script_confusables)] // This is to allow structs created by a derive macro to have private fields that begin with the grek letter π
 
 pub mod artifact_service;
+pub mod build_service;
 pub mod cli_commands;
 pub mod docker;
 pub mod java;
 pub mod logging;
 pub mod network;
 pub mod node_api;
+pub mod peer_metrics;
 pub mod transparency_log;
 pub mod util;
