+++ src/node_api/model/cli.rs
@@ -15,34 +15,15 @@
 */
 
 use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct Status {
     pub peers_count: usize,
     pub peer_id: String,
-    pub artifact_count: ArtifactsSummary,
-    pub disk_allocated: String,
-    pub disk_usage: String,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct ArtifactsSummary {
-    pub total: String,
-    pub summary: HashMap<String, usize>,
 }
 
 impl std::fmt::Display for Status {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        writeln!(f, "Connected Peers Count:       {}", self.peers_count)?;
-        writeln!(f, "Artifacts Count:             {}", self.artifact_count)?;
-        writeln!(f, "Total Disk Space Allocated:  {}", self.disk_allocated)?;
-        write!(f, "Disk Space Used:             {}%", self.disk_usage)
-    }
-}
-
-impl std::fmt::Display for ArtifactsSummary {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "{} {:?}", self.total, self.summary)
+        write!(f, "Connected Peers Count:       {}", self.peers_count)
     }
 }
