+++ src/peer_metrics.rs
@@ -14,6 +14,4 @@
    limitations under the License.
 */
 
-pub mod hashing;
-pub mod service;
-pub mod storage;
+pub mod metrics;
